python-services/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
keybert
sentence-transformers
psycopg2-binary
python-dotenv
apscheduler==3.10.4
httpx
pytest
